requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
pdfminer.six>=20231228,<20260000
google-generativeai==0.3.2
beautifulsoup4==4.12.2
requests==2.31.0
